+++ blueprints/model/index.js
@@ -9,7 +9,7 @@ module.exports = Blueprint.extend({
     var entityOptions = options.entity.options;
 
     for (var name in entityOptions) {
-      var type = entityOptions[name];
+      var type = entityOptions[name] || '';
       var dasherizedName = stringUtils.dasherize(name);
       var dasherizedNameSingular = inflection.singularize(dasherizedName);
       var camelizedName = stringUtils.camelize(name);
@@ -51,5 +51,9 @@ function dsAttr(name, type) {
   case 'has-many':
     var singularizedName = inflection.singularize(name);
     return 'DS.hasMany(\'' + singularizedName + '\')';
+  default:
+    //"If you don't specify the type of the attribute, it will be whatever was provided by the server"
+    //http://emberjs.com/guides/models/defining-models/
+    return 'DS.attr()';
   }
 }